% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_decompose.R
\name{plot.hlm_decompose}
\alias{plot.hlm_decompose}
\title{Plot method for hlm_decompose objects}
\usage{
\method{plot}{hlm_decompose}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hlm_decompose}.}

\item{...}{Not used.}
}
\value{
A ggplot object.
}
\description{
Produces a simple bar chart of variance shares across components,
suitable for teaching how variance is partitioned across levels.
}
