% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_context.R
\name{hlm_context_plot}
\alias{hlm_context_plot}
\title{Convenience wrapper to plot contextual effects}
\usage{
hlm_context_plot(object)
}
\arguments{
\item{object}{An object of class \code{hlm_context}.}
}
\value{
A ggplot object.
}
\description{
Convenience wrapper to plot contextual effects
}
